Refactor types

+++ src/sql/adapter/mod.rs
@@ -1,12 +1,12 @@
 pub trait SqlAdapter {
-    fn placeholder(&self, idx: usize) -> String;
+    fn placeholder(&self, idx: u8) -> String;
 }
 
 #[derive(Copy)]
 pub struct PostgreSqlAdapter;
 
 impl SqlAdapter for PostgreSqlAdapter {
-    fn placeholder(&self, idx: usize) -> String {
+    fn placeholder(&self, idx: u8) -> String {
         format!("${}", idx)
     }
 }
